Items from the list are removable

+++ app/src/main/java/com/example/nutri/domain/model/Ingredient.kt
@@ -2,14 +2,14 @@ package com.example.nutri.domain.model
 
 class Ingredient (
     val ingredientName : String,
-    val ingredientAmount: Double,
+    val ingredientAmount: Int,
     val ingredientUnits: String
 ){
 
     companion object{
         fun emptyIngredient() = Ingredient(
             ingredientName = "",
-            ingredientAmount = 0.0,
+            ingredientAmount = 0,
             ingredientUnits = "g"
         )
     }
